Re-engineer the CMake system
The CMake build behavior now more closely aligns with the original
behavior, where the user has to provide a CMake.system to set up all the
paths to the dependencies etc.

However, a more automated behavior can be enabled by passing certain
configuration options to CMake when configuring the build.